+++ tiledesk-chatbot-util/test/test_split.js
@@ -56,11 +56,11 @@ describe('TiledeskChatbotUtil', function() {
             assert.strictEqual(commands[0].type, TiledeskChatbotUtil.COMMAND_TYPE_MESSAGE);
             assert.strictEqual(commands[0].text, 'Intro text');
             assert.strictEqual(commands[1].type, TiledeskChatbotUtil.COMMAND_TYPE_WAIT);
-            assert.strictEqual(commands[1].time, 1000);
+            assert.strictEqual(commands[1].time, 500);
             assert.strictEqual(commands[2].type, TiledeskChatbotUtil.COMMAND_TYPE_MESSAGE);
             assert.strictEqual(commands[2].text, 'second message');
             assert.strictEqual(commands[3].type, TiledeskChatbotUtil.COMMAND_TYPE_WAIT);
-            assert.strictEqual(commands[3].time, 1000);
+            assert.strictEqual(commands[3].time, 500);
             assert.strictEqual(commands[4].type, TiledeskChatbotUtil.COMMAND_TYPE_MESSAGE);
             assert.strictEqual(commands[4].text, 'third message');
         });
@@ -78,13 +78,26 @@ describe('TiledeskChatbotUtil', function() {
             assert(commands.length == 4);
             
             assert.strictEqual(commands[0].type, TiledeskChatbotUtil.COMMAND_TYPE_WAIT);
-            assert.strictEqual(commands[0].time, 1000);
+            assert.strictEqual(commands[0].time, 500);
             assert.strictEqual(commands[1].type, TiledeskChatbotUtil.COMMAND_TYPE_MESSAGE);
             assert.strictEqual(commands[1].text, 'Intro text');
             assert.strictEqual(commands[2].type, TiledeskChatbotUtil.COMMAND_TYPE_WAIT);
-            assert.strictEqual(commands[2].time, 1000);
+            assert.strictEqual(commands[2].time, 500);
             assert.strictEqual(commands[3].type, TiledeskChatbotUtil.COMMAND_TYPE_MESSAGE);
             assert.strictEqual(commands[3].text, 'second message');
         });
     });
 });
+
+describe('TiledeskChatbotUtil', function() {
+    describe('splitMessage()', function() {
+        it('should return NO commands', function() {
+            const text = "One message";
+            console.log("splitting text:", text);
+            var commands = TiledeskChatbotUtil.splitMessage(text);
+            console.log("Commands:", commands)
+            assert(commands == null);
+        });
+    });
+});
+