+++ tiledesk-chatbot-util/test/test_split.sh
@@ -1,0 +1,1 @@
+npx mocha ./test/test_split.js --exit