+++ tiledesk-chatbot-util/test/test_split_pars.sh
@@ -1,0 +1,1 @@
+npx mocha ./test/test_split_pars.js --exit